🐛 FIX: Replace the buggy shields bagde
Thanks for the awesome theme Sarah. I've also authored a [purple VSCode theme](https://marketplace.visualstudio.com/items?itemName=ahmadawais.shades-of-purple). The shields badge for downloads which was working alright is now buggy. It was reporting fewer downloads than actual so I removed it.

I first thought it was only for my theme but apparently, it's reporting 20k fewer downloads for Night Owl as well. So, I figured you'd want to remove it.

I replaced it with the reviews badge.